Struct rccheck::Psk

pub struct Psk { /* private fields */ }

The self-referencing struct.

Implementations

impl Psk

pub fn with_key_bytes<'outer_borrow, ReturnType>(
&'outer_borrow self,
user: impl for<'this> FnOnce(&'outer_borrow Vec<u8>) -> ReturnType
) -> ReturnType

Provides limited immutable access to key_bytes. This method was generated because the contents of key_bytes are immutably borrowed by other fields.

pub fn borrow_key_bytes<'this>(&'this self) -> &'this Vec<u8>

Provides limited immutable access to key_bytes. This method was generated because the contents of key_bytes are immutably borrowed by other fields.

pub fn with_spki<'outer_borrow, ReturnType>(
&'outer_borrow self,
user: impl for<'this> FnOnce(&'outer_borrow SubjectPublicKeyInfo<'this>) -> ReturnType
) -> ReturnType

Provides an immutable reference to spki. This method was generated because spki is a tail field.

pub fn borrow_spki<'this>(&'this self) -> &'this SubjectPublicKeyInfo<'this>

Provides an immutable reference to spki. This method was generated because spki is a tail field.

pub fn with_spki_mut<'outer_borrow, ReturnType>(
&'outer_borrow mut self,
user: impl for<'this> FnOnce(&'outer_borrow mut SubjectPublicKeyInfo<'this>) -> ReturnType
) -> ReturnType

Provides a mutable reference to spki. This method was generated because spki is a tail field. No borrow_spki_mut function was generated because Rust’s borrow checker is currently unable to guarantee that such a method would be used safely.

impl Psk

pub fn from_der(bytes: &[u8]) -> Result<Psk, Error>